+++ tasks.py
@@ -43,7 +43,11 @@ def coverage(ctx):
 def serve_coverage(ctx, host='localhost', port=8000):
     """Generate coverage report and use Python's built-in HTTP server to
     serve it locally."""
-    ctx.run('python -m http.server {} -d reports/coverage/html -b {}'.format(port, host))
+    ctx.run(
+        'python -m http.server {} -d reports/coverage/html -b {}'.format(
+            port, host
+        )
+    )
 
 
 @invoke.task
@@ -149,7 +153,9 @@ def build(_):
 def serve_docs(ctx, host='localhost', port=8000):
     """Generate documentation and use Python's built-in HTTP server to serve
     it locally."""
-    ctx.run('python -m http.server {} -d docs/build/html -b {}'.format(port, host))
+    ctx.run(
+        'python -m http.server {} -d docs/build/html -b {}'.format(port, host)
+    )
 
 
 @invoke.task
@@ -189,7 +195,7 @@ def clean(ctx):
     """Clean working directory."""
     ctx.run('find . -name "*.pyc" -delete')
     ctx.run('find . -type d -name "__pycache__" -empty -delete')
-    ctx.run('rm mockify/_version.py')
+    ctx.run('rm -rf mockify/_version.py')
     ctx.run('rm -rf docs/build')
     ctx.run('rm -rf build dist')
     ctx.run('rm -rf *.egg-info')