+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
   * New action added: :class:`mockify.actions.ReturnAsync`
   * New action added: :class:`mockify.actions.IterateAsync`
+  * New action added: :class:`mockify.actions.RaiseAsync`
 
 **Changed**